requirements.txt
accelerate==1.11.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bert-score==0.3.13
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
colorlog==6.10.1
contourpy==1.3.3
cycler==0.12.1
datasets==4.4.1
dill==0.4.0
distro==1.9.0
entmax==1.3
evaluate==0.4.6
filelock==3.19.1
filetype==1.2.0
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.2.1
google-genai==1.59.0
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
graphviz==0.21
greenlet==3.2.4
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
ipadic==1.0.0
Jinja2==3.1.6
joblib==1.5.2
jsonargparse==3.13.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain-core==1.2.7
langchain-google-genai==4.2.0
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
lightning-utilities==0.15.2
lxml==6.0.2
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.10.7
mecab-python3==1.0.12
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
narwhals==2.11.0
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
optuna==4.6.0
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.4.0
portalocker==3.2.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.3
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.29.0
rsa==4.9.1
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.3
seaborn==0.13.2
sentencepiece==0.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.44
streamlit==1.51.0
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.1
toml==0.10.2
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchmetrics==0.10.3
torchvision==0.22.1+cu118
tornado==6.5.2
tqdm==4.67.1
transformers==4.57.1
triton==3.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unbabel-comet==2.2.7
uritemplate==4.2.0
urllib3==2.5.0
uuid_utils==0.13.0
watchdog==6.0.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
